requirements.txt
appnope==0.1.0
blessed==1.14.2
botocore==1.10.28
cairocffi==0.8.0
CairoSVG==2.1.3
cement==2.8.2
cffi==1.11.5
cssselect2==0.2.1
decorator==4.0.11
defusedxml==0.5.0
Django==1.10.5
django-admin-views==0.8.0
docker==2.6.1
docker-pycreds==0.2.3
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
html5lib==1.0.1
ipython==5.3.0
ipython-genutils==0.1.0
Jinja2==2.9.4
jmespath==0.9.3
MarkupSafe==0.23
numpy==1.12.0
pandas==0.19.2
pathspec==0.5.5
pdfrw==0.4
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pycparser==2.18
Pygments==2.2.0
Pyphen==0.9.4
python-dateutil==2.7.3
pytz==2016.10
PyYAML==3.12
quantum-ql==0.0.1
requests==2.13.0
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.10.0
termcolor==1.1.0
tinycss2==0.6.1
traitlets==4.3.2
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
websocket-client==0.48.0
